requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-filter==21.1
django-model-utils==4.2.0
Pillow==8.4.0
psycopg2==2.9.2
python-dotenv==0.19.2
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tqdm==4.62.3
